Connection/BLood Donation Api/requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
gunicorn==21.2.0
joblib==1.3.2
numpy==1.26.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
